requirements.txt
# Работа с HTTP-запросами к API OZON
requests==2.32.3

# Работа с PostgreSQL
psycopg2-binary==2.9.10

# Удобная работа с переменными окружения (.env файлы)
python-dotenv==1.0.1

# (опционально, потом может пригодиться для анализа/отладки)
pandas==2.2.3